Make large integer support conditional
